WIP add custom function scopes
Custom function scopes take advantage of the fact that blocks have a return type of `void`,
and that function arguments are evaluated before being passed to the body. These scopes
can be used to cleanly implement open/close patterns, cutting down on LOC and
improving readability.

That said, a major drawback is that the "close" function (e.g. `boxEnd()`) will not be called
if control flow prematurely exits the scope. You can fall back to the normal API in that case though.